refactor(autoware_agnocast_wrapper): move header contents into per-topic files
