+++ common/autoware_agnocast_wrapper/include/autoware/agnocast_wrapper/subscription.hpp
@@ -1,0 +1,163 @@
+// Copyright 2025 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifdef USE_AGNOCAST_ENABLED
+
+namespace autoware::agnocast_wrapper
+{
+
+template <typename MessageT>
+class Subscription
+{
+public:
+  using SharedPtr = std::shared_ptr<Subscription<MessageT>>;
+
+  virtual ~Subscription() = default;
+};
+
+template <typename MessageT>
+class AgnocastSubscription : public Subscription<MessageT>
+{
+  typename agnocast::Subscription<MessageT>::SharedPtr subscription_;
+
+public:
+  template <typename NodeT, typename Func>
+  explicit AgnocastSubscription(
+    NodeT * node, const std::string & topic_name, const rclcpp::QoS & qos, Func && callback,
+    const agnocast::SubscriptionOptions & options)
+  {
+    // TODO(Koichi98): AUTOWARE_MESSAGE_UNIQUE_PTR should be disallowed for Agnocast subscriptions.
+    // Agnocast uses shared memory, so mutable exclusive ownership is semantically incorrect and
+    // risks corrupting data read by other subscribers. Currently kept for compatibility with
+    // CudaPointcloudPreprocessorNode which uses UNIQUE_PTR callbacks.
+    static_assert(
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_UNIQUE_PTR(MessageT) &&> ||
+        std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_CONST_SHARED_PTR(MessageT) &&> ||
+        std::is_invocable_v<std::decay_t<Func>, const MessageT &>,
+      "callback should be invocable with an rvalue reference to either "
+      "AUTOWARE_MESSAGE_UNIQUE_PTR or AUTOWARE_MESSAGE_CONST_SHARED_PTR, or with a "
+      "const reference to the message type");
+
+    constexpr bool is_message_ptr_callback =
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_UNIQUE_PTR(MessageT) &&> ||
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_CONST_SHARED_PTR(MessageT) &&>;
+    constexpr auto ownership =
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_UNIQUE_PTR(MessageT) &&>
+        ? OwnershipType::Unique
+        : OwnershipType::Shared;
+
+    subscription_ = agnocast::create_subscription<MessageT>(
+      node, topic_name, qos,
+      [callback = std::forward<Func>(callback)](agnocast::ipc_shared_ptr<MessageT> && msg) {
+        if constexpr (!is_message_ptr_callback) {
+          // msg keeps the shared-memory entry alive only while the callback runs: the
+          // reference is valid for the duration of the callback and no copy is made, but
+          // it must not be stored or used after the callback returns. Callbacks that need
+          // to extend the message lifetime should take AUTOWARE_MESSAGE_CONST_SHARED_PTR.
+          // as_const prevents generic callbacks from mutating the shared-memory entry,
+          // which other processes may be reading concurrently.
+          callback(std::as_const(*msg));
+        } else if constexpr (ownership == OwnershipType::Unique) {
+          callback(message_ptr<MessageT, ownership>(std::move(msg)));
+        } else {
+          callback(
+            message_ptr<const MessageT, ownership>(
+              agnocast::ipc_shared_ptr<const MessageT>(std::move(msg))));
+        }
+      },
+      options);
+  }
+};
+
+template <typename MessageT>
+class ROS2Subscription : public Subscription<MessageT>
+{
+  typename rclcpp::Subscription<MessageT>::SharedPtr subscription_;
+
+public:
+  template <typename Func>
+  explicit ROS2Subscription(
+    rclcpp::Node * node, const std::string & topic_name, const rclcpp::QoS & qos, Func && callback,
+    const agnocast::SubscriptionOptions & options)
+  {
+    static_assert(
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_UNIQUE_PTR(MessageT) &&> ||
+        std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_CONST_SHARED_PTR(MessageT) &&> ||
+        std::is_invocable_v<std::decay_t<Func>, const MessageT &>,
+      "callback should be invocable with an rvalue reference to either "
+      "AUTOWARE_MESSAGE_UNIQUE_PTR or AUTOWARE_MESSAGE_CONST_SHARED_PTR, or with a "
+      "const reference to the message type");
+
+    constexpr bool is_message_ptr_callback =
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_UNIQUE_PTR(MessageT) &&> ||
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_CONST_SHARED_PTR(MessageT) &&>;
+    constexpr auto ownership =
+      std::is_invocable_v<std::decay_t<Func>, AUTOWARE_MESSAGE_UNIQUE_PTR(MessageT) &&>
+        ? OwnershipType::Unique
+        : OwnershipType::Shared;
+
+    rclcpp::SubscriptionOptions ros2_options;
+    ros2_options.callback_group = options.callback_group;
+    subscription_ = node->create_subscription<MessageT>(
+      topic_name, qos,
+      [callback = std::forward<Func>(callback)](std::unique_ptr<MessageT> msg) {
+        if constexpr (!is_message_ptr_callback) {
+          // as_const keeps this fallback consistent with the Agnocast path: generic
+          // callbacks must not observe a mutable reference on either path.
+          callback(std::as_const(*msg));
+        } else if constexpr (ownership == OwnershipType::Unique) {
+          callback(message_ptr<MessageT, ownership>(std::move(msg)));
+        } else {
+          callback(
+            message_ptr<const MessageT, ownership>(
+              std::shared_ptr<const MessageT>(std::move(msg))));
+        }
+      },
+      ros2_options);
+  }
+};
+
+template <typename MessageT, typename Func>
+typename Subscription<MessageT>::SharedPtr create_subscription(
+  rclcpp::Node * node, const std::string & topic_name, const rclcpp::QoS & qos, Func && callback,
+  const agnocast::SubscriptionOptions & options)
+{
+  if (use_agnocast()) {
+    return std::make_shared<AgnocastSubscription<MessageT>>(
+      node, topic_name, qos, std::forward<Func>(callback), options);
+  } else {
+    return std::make_shared<ROS2Subscription<MessageT>>(
+      node, topic_name, qos, std::forward<Func>(callback), options);
+  }
+}
+
+template <typename MessageT, typename Func>
+typename Subscription<MessageT>::SharedPtr create_subscription(
+  rclcpp::Node * node, const std::string & topic_name, const size_t qos_history_depth,
+  Func && callback, const agnocast::SubscriptionOptions & options)
+{
+  if (use_agnocast()) {
+    return std::make_shared<AgnocastSubscription<MessageT>>(
+      node, topic_name, rclcpp::QoS(rclcpp::KeepLast(qos_history_depth)),
+      std::forward<Func>(callback), options);
+  } else {
+    return std::make_shared<ROS2Subscription<MessageT>>(
+      node, topic_name, rclcpp::QoS(rclcpp::KeepLast(qos_history_depth)),
+      std::forward<Func>(callback), options);
+  }
+}
+
+}  // namespace autoware::agnocast_wrapper
+
+#endif
